Fix patch script import chain so it runs without scipy on ziggy.
Move summary discovery helpers to darkhunter_rv/summary_paths.py so lightweight
scripts avoid importing fit_apf_rv_keplerian.
